refactor(session-model): extract selection and configuration owners
Move Session runtime selection out of the workspace runtime facade and Session
agent configuration out of WorkspacePage. Raise the runtime owner gate to 700
and the page gate to 1230 so per-session target wiring is not compressed for
line count.
